test: Add comprehensive autorouting error handling tests for issue #763
Reproduce and validate fix for cryptic "Unexpected numItems value: 0" errors
when components overlap, ensuring users get actionable error messages.
