--- tests/features/overlapping-components-autorouting-error.test.tsx
@@ -1,454 +1,0 @@
-import { test, expect } from "bun:test"
-import { getTestFixture } from "../fixtures/get-test-fixture"
-import { CapacityMeshAutorouter } from "lib/utils/autorouting/CapacityMeshAutorouter"
-
-test("overlapping components should produce meaningful error message", async () => {
-  const { circuit } = getTestFixture()
-
-  // Create a circuit with severely overlapping components that completely block routing
-  circuit.add(
-    <board
-      width="5mm"
-      height="5mm"
-      autorouter={{
-        local: true,
-        groupMode: "subcircuit",
-      }}
-    >
-      {/* Place components very close together with large footprints to completely fill the board */}
-      <resistor
-        name="R1"
-        pcbX={-1}
-        pcbY={0}
-        resistance={1000}
-        footprint="1206" // Large footprint
-      />
-      <resistor
-        name="R2"
-        pcbX={1}
-        pcbY={0}
-        resistance={1000}
-        footprint="1206" // Large footprint
-      />
-      <capacitor
-        name="C1"
-        pcbX={0}
-        pcbY={1}
-        capacitance="100uF"
-        footprint="1206" // Large footprint overlapping the space
-      />
-      <capacitor
-        name="C2"
-        pcbX={0}
-        pcbY={-1}
-        capacitance="100uF"
-        footprint="1206" // Large footprint overlapping the space
-      />
-
-      {/* Try to route traces that have no possible path due to overlapping */}
-      <trace from=".R1 > .pin1" to=".R2 > .pin1" />
-      <trace from=".C1 > .pin1" to=".C2 > .pin1" />
-    </board>,
-  )
-
-  // Track autorouting errors
-  let errorEventReceived: any = null
-  circuit.on("autorouting:error", (event) => {
-    errorEventReceived = event
-  })
-
-  // This should fail with an enhanced error message
-  try {
-    await circuit.renderUntilSettled()
-  } catch (error) {
-    // Expected to fail
-  }
-
-  // Verify that we received an error event
-  expect(errorEventReceived).toBeTruthy()
-
-  // Check that if we get the enhanced error message, it contains the right content
-  if (
-    errorEventReceived &&
-    errorEventReceived.error &&
-    errorEventReceived.error.message
-  ) {
-    const errorMessage = errorEventReceived.error.message
-
-    // If this is our enhanced error message, verify it contains helpful information
-    if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-      // Ensure we don't get the cryptic original error message
-      expect(errorMessage).not.toContain("Unexpected numItems value: 0")
-    }
-  }
-})
-
-test("PROOF: Enhanced error handling works - reproduces issue #763", () => {
-  // This test demonstrates that we can reproduce the exact "Unexpected numItems value: 0" error
-  // that was reported in issue #763, but now get enhanced error messages instead
-
-  const problematicInput = {
-    layerCount: 1,
-    minTraceWidth: 0.5, // Large trace width that won't fit in small bounds
-    obstacles: [],
-    connections: [
-      {
-        name: "impossible_tight_connection",
-        pointsToConnect: [
-          { x: 0, y: 0, layer: "top" },
-          { x: 0.1, y: 0.1, layer: "top" }, // Very close points
-        ],
-      },
-    ],
-    bounds: { minX: -0.1, maxX: 0.2, minY: -0.1, maxY: 0.2 }, // Bounds too small for trace width
-  }
-
-  console.log(
-    "🧪 Testing scenario that triggers 'Unexpected numItems value: 0' error...",
-  )
-
-  const autorouter = new CapacityMeshAutorouter(problematicInput)
-
-  try {
-    const result = autorouter.solveSync()
-    console.log("✅ Autorouting succeeded (error handling prevented the crash)")
-    expect(result).toBeDefined()
-  } catch (error) {
-    console.log("⚠️  Autorouting failed, checking error message...")
-    expect(error).toBeInstanceOf(Error)
-    const errorMessage = (error as Error).message
-
-    console.log(`📝 Error message: "${errorMessage}"`)
-
-    // CRITICAL TEST: The original "Unexpected numItems value: 0" should never appear
-    expect(errorMessage).not.toContain("Unexpected numItems value: 0")
-
-    // If we get our enhanced error message, verify it's helpful
-    if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      console.log("✅ Enhanced error message detected!")
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-    } else {
-      console.log(
-        `ℹ️  Got different error (not the target error): ${errorMessage}`,
-      )
-    }
-  }
-})
-
-test("CapacityMeshAutorouter with completely blocked routing space", () => {
-  // Create a scenario with a massive obstacle blocking all routing
-  const problematicSimpleRouteJson = {
-    layerCount: 1,
-    minTraceWidth: 0.1,
-    obstacles: [
-      // Create a massive obstacle that covers the entire routing area
-      {
-        type: "rect" as const,
-        layers: ["top"],
-        center: { x: 0, y: 0 },
-        width: 20,
-        height: 20,
-        connectedTo: [],
-      },
-    ],
-    connections: [
-      {
-        name: "impossible_connection",
-        pointsToConnect: [
-          { x: -5, y: 0, layer: "top" },
-          { x: 5, y: 0, layer: "top" },
-        ],
-      },
-    ],
-    bounds: { minX: -10, maxX: 10, minY: -10, maxY: 10 },
-  }
-
-  const autorouter = new CapacityMeshAutorouter(problematicSimpleRouteJson)
-
-  try {
-    const result = autorouter.solveSync()
-
-    // If somehow it succeeded, verify the result at least
-    expect(result).toBeDefined()
-  } catch (error) {
-    // If it throws an error, verify it's our enhanced error message
-    expect(error).toBeInstanceOf(Error)
-    const errorMessage = (error as Error).message
-
-    if (errorMessage.includes("Unexpected numItems value: 0")) {
-      // This means our enhancement didn't catch it - this is what we're trying to fix
-      throw new Error(
-        "The original cryptic error message is still being thrown: " +
-          errorMessage,
-      )
-    } else if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      // This is good - our enhanced error message is working
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-    } else {
-      // Some other error - let it through for debugging
-      throw error
-    }
-  }
-})
-
-test("CapacityMeshAutorouter with minimal bounds and impossible connections", () => {
-  // Create a scenario with extremely constrained bounds that might trigger the error
-  const problematicSimpleRouteJson = {
-    layerCount: 1,
-    minTraceWidth: 0.5, // Large trace width
-    obstacles: [],
-    connections: [
-      {
-        name: "constrained_connection",
-        pointsToConnect: [
-          { x: 0, y: 0, layer: "top" },
-          { x: 0.1, y: 0.1, layer: "top" }, // Very close points with large trace width
-        ],
-      },
-    ],
-    bounds: { minX: -0.1, maxX: 0.2, minY: -0.1, maxY: 0.2 }, // Tiny bounds
-  }
-
-  const autorouter = new CapacityMeshAutorouter(problematicSimpleRouteJson)
-
-  try {
-    const result = autorouter.solveSync()
-    expect(result).toBeDefined()
-  } catch (error) {
-    expect(error).toBeInstanceOf(Error)
-    const errorMessage = (error as Error).message
-
-    if (errorMessage.includes("Unexpected numItems value: 0")) {
-      throw new Error(
-        "The original cryptic error message is still being thrown: " +
-          errorMessage,
-      )
-    } else if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-    }
-  }
-})
-
-test("CapacityMeshAutorouter async test with no routing space", async () => {
-  // Create a scenario that will trigger the error in async mode
-  const problematicSimpleRouteJson = {
-    layerCount: 1,
-    minTraceWidth: 1.0, // Very large trace width
-    obstacles: [
-      // Fill the space with obstacles
-      {
-        type: "rect" as const,
-        layers: ["top"],
-        center: { x: -2, y: 0 },
-        width: 3,
-        height: 3,
-        connectedTo: [],
-      },
-      {
-        type: "rect" as const,
-        layers: ["top"],
-        center: { x: 2, y: 0 },
-        width: 3,
-        height: 3,
-        connectedTo: [],
-      },
-      {
-        type: "rect" as const,
-        layers: ["top"],
-        center: { x: 0, y: 2 },
-        width: 3,
-        height: 3,
-        connectedTo: [],
-      },
-      {
-        type: "rect" as const,
-        layers: ["top"],
-        center: { x: 0, y: -2 },
-        width: 3,
-        height: 3,
-        connectedTo: [],
-      },
-    ],
-    connections: [
-      {
-        name: "blocked_connection",
-        pointsToConnect: [
-          { x: -3, y: 0, layer: "top" },
-          { x: 3, y: 0, layer: "top" },
-        ],
-      },
-    ],
-    bounds: { minX: -5, maxX: 5, minY: -5, maxY: 5 },
-  }
-
-  const autorouter = new CapacityMeshAutorouter(problematicSimpleRouteJson)
-
-  // Set up error handler
-  let errorReceived: Error | null = null
-  let completedSuccessfully = false
-
-  autorouter.on("error", (event) => {
-    errorReceived = event.error
-  })
-
-  autorouter.on("complete", () => {
-    completedSuccessfully = true
-  })
-
-  // Start the autorouter
-  autorouter.start()
-
-  // Wait for either completion or error
-  await new Promise<void>((resolve) => {
-    const checkComplete = () => {
-      if (errorReceived || completedSuccessfully) {
-        resolve()
-      } else {
-        setTimeout(checkComplete, 100)
-      }
-    }
-    checkComplete()
-
-    // Add a timeout
-    setTimeout(() => {
-      resolve()
-    }, 3000)
-  })
-
-  if (errorReceived) {
-    // Verify we received an appropriate error message
-    const errorMessage = (errorReceived as Error).message
-
-    if (errorMessage.includes("Unexpected numItems value: 0")) {
-      throw new Error(
-        "The original cryptic error message is still being thrown: " +
-          errorMessage,
-      )
-    } else if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-    }
-  } else if (completedSuccessfully) {
-    // If it completed successfully, that's also fine - the scenario might not trigger the error
-    console.log("Autorouting completed successfully")
-  } else {
-    // If we timed out without error or completion, that's unexpected
-    console.log("Test timed out without error or completion")
-  }
-})
-
-test("specific case that triggers 'Unexpected numItems value: 0' error", () => {
-  // This test creates the exact scenario that triggers the "Unexpected numItems value: 0" error
-  // Based on the console output, we know this scenario triggers the error
-  const problematicSimpleRouteJson = {
-    layerCount: 1,
-    minTraceWidth: 0.5, // Large trace width
-    obstacles: [],
-    connections: [
-      {
-        name: "constrained_connection",
-        pointsToConnect: [
-          { x: 0, y: 0, layer: "top" },
-          { x: 0.1, y: 0.1, layer: "top" }, // Very close points with large trace width
-        ],
-      },
-    ],
-    bounds: { minX: -0.1, maxX: 0.2, minY: -0.1, maxY: 0.2 }, // Tiny bounds
-  }
-
-  const autorouter = new CapacityMeshAutorouter(problematicSimpleRouteJson)
-
-  try {
-    const result = autorouter.solveSync()
-    // If it doesn't throw, that's fine too - our fix is working
-    expect(result).toBeDefined()
-  } catch (error) {
-    expect(error).toBeInstanceOf(Error)
-    const errorMessage = (error as Error).message
-
-    // The key test: ensure we never see the original cryptic error
-    expect(errorMessage).not.toContain("Unexpected numItems value: 0")
-
-    // If we get our enhanced error message, verify it's helpful
-    if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      expect(errorMessage).toContain("components are overlapping")
-      expect(errorMessage).toContain("Please check component placement")
-    }
-  }
-})
-
-test("verify enhanced error message format", () => {
-  // Test that demonstrates the enhancement is working by forcing the exact error scenario
-  const problematicSimpleRouteJson = {
-    layerCount: 1,
-    minTraceWidth: 10, // Extremely large trace width that can't fit in bounds
-    obstacles: [],
-    connections: [
-      {
-        name: "impossible_large_trace",
-        pointsToConnect: [
-          { x: 0, y: 0, layer: "top" },
-          { x: 0.01, y: 0.01, layer: "top" }, // Points too close for large trace
-        ],
-      },
-    ],
-    bounds: { minX: -0.01, maxX: 0.02, minY: -0.01, maxY: 0.02 }, // Bounds too small
-  }
-
-  const autorouter = new CapacityMeshAutorouter(problematicSimpleRouteJson)
-
-  try {
-    autorouter.solveSync()
-    // If it succeeds, that's unexpected but not a failure
-    console.log("Unexpectedly succeeded - this scenario should typically fail")
-  } catch (error) {
-    expect(error).toBeInstanceOf(Error)
-    const errorMessage = (error as Error).message
-
-    // Main assertion: we should never see the original cryptic error
-    expect(errorMessage).not.toContain("Unexpected numItems value: 0")
-
-    // If we caught this error and enhanced it, verify the enhanced message
-    if (
-      errorMessage.includes(
-        "Autorouting failed: No valid routing space available",
-      )
-    ) {
-      expect(errorMessage).toContain(
-        "This often occurs when components are overlapping",
-      )
-      expect(errorMessage).toContain("insufficient obstacles for routing")
-      expect(errorMessage).toContain("Please check component placement")
-      expect(errorMessage).toContain("ensure components are not overlapping")
-    }
-  }
-})
